requirements.txt
fastapi>=0.116.2
uvicorn>=0.27.1
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
qdrant-client>=1.8.0
neo4j>=5.18.0
langgraph>=0.0.26
langchain-community>=0.0.25
langchain-core>=0.1.28
langdetect>=1.0.9
requests>=2.31.0
python-multipart>=0.0.9
torch>=2.2.1
torchvision>=0.17.1
torchaudio>=2.2.1
transformers>=4.38.2
sentencepiece>=0.2.0
sacremoses>=0.1.1
redis>=5.0.1
google-api-python-client>=2.122.0
google-auth>=2.28.2
duckduckgo-search>=8.1.1
pillow>=10.2.0
numpy>=1.26.4
faster-whisper>=1.0.1
librosa>=0.10.1
scipy>=1.12.0
cognee[neo4j,qdrant]>=1.0.0
cognee-community-vector-adapter-qdrant
groq
